Käyttäjän luonti tehty
Ikkuna käyttäjän luomiselle tehty. Lomake tarkistaa onko käyttäjä tunnus ja varattu, ilmoittaa käyttäjälle jos on, eikä anna luoda uutta samanlaista käyttäjää. Lisäksi salasana ja salasana uudestaan tarkistetaan, että ovat saman.
Kuvan lisäämistä ei vielä implementoitu, koska tarvitsee oman endpointin, eikä siitä saa lisäpisteitä.
