+++ MoviePlanet/client/src/createcustomer.css
@@ -1,15 +1,15 @@
-/* Styling modal window */
+/* Styling modal window*/
 .modal {
     position: fixed;
     top: 50%; 
     left: 50%; 
-    transform: translate(-50%, -50%);
-    width: 60%;
-    max-width: 400px;
+    transform: translate(-50%, -50%); 
+    width: 60%; 
+    max-width: 400px; 
     background-color: white;
     padding: 20px;
     border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.3); 
   }
 
   .close {
@@ -21,7 +21,7 @@
     color: #888;
   }
 
-  #LoginFormLoginBtn, input[type=text], input[type=password] {
+  #RegisterBtn, input[type=text], input[type=password] {
     display: block;
     width: calc(100% - 20px);
     margin-bottom: 10px;
@@ -31,12 +31,17 @@
     font-size: 16px;
   }
 
-  #LoginFormLoginBtn {
+  #RegisterBtn {
     cursor: pointer;
   }
 
   .error {
     color: red;
+    font-size: 14px;
+    margin-top: 5px;
+  }
+  .success {
+    color: green;
     font-size: 14px;
     margin-top: 5px;
   }
